+++ extensions/resource-inspector/eslint.config.js
@@ -1,2 +1,6 @@
+const { defineConfig } = require("eslint/config");
 const raycast = require("@raycast/eslint-config");
-module.exports = [...raycast, { ignores: ["dist/**", "raycast-env.d.ts"] }];
+module.exports = defineConfig([
+  ...raycast,
+  { ignores: ["dist/**", "raycast-env.d.ts"] },
+]);